requirements.txt
pytube
openai-whisper
transformers
torch
gradio